requirements.txt
streamlit==1.25.0
torch>=1.10.0
transformers
Pillow
easyocr
